Fold .hpp files in framework/include/cppmicroservices/detail into their corresponding .h files
