+++ framework/include/CMakeLists.txt
@@ -37,11 +37,8 @@ set(_public_headers
   cppmicroservices/ServiceRegistrationBase.h
   cppmicroservices/ServiceTracker.h
   cppmicroservices/ServiceTrackerCustomizer.h
-  cppmicroservices/detail/ServiceTracker.hpp
   cppmicroservices/detail/ServiceTrackerPrivate.h
-  cppmicroservices/detail/ServiceTrackerPrivate.hpp
   cppmicroservices/detail/TrackedService.h
-  cppmicroservices/detail/TrackedService.hpp
   cppmicroservices/detail/TrackedServiceListener.h
 
   cppmicroservices/Bundle.h
@@ -58,7 +55,6 @@ set(_public_headers
   cppmicroservices/Constants.h
   cppmicroservices/GetBundleContext.h
   cppmicroservices/detail/BundleAbstractTracked.h
-  cppmicroservices/detail/BundleAbstractTracked.hpp
   cppmicroservices/detail/BundleResourceBuffer.h
   cppmicroservices/detail/ScopeGuard.h
   cppmicroservices/detail/CounterLatch.h
